Add conditions_dict to accumulate train/inference conditions from transforms (#272)
With everything after `conf.transforms` supposed to be constant in
train/inference. It follows that the extra_tXd loaders are no longer the
place to do complex thinking about the conditions (as test/train would
obviously think differently).

This PR primarily adds a new dictionary variable called
`conditions_dict` to the transforms that's initialized at the dataloader
and then passed all the way through to `**conditions_dict` in
get_extra_tXd.

Additionally, there's a new Transform to update `conditions_dict` after
transform indep modifies the indep. (Since the length changes). I figure
a transform to handle this is better than each extra_tXd method
adjusting for the new length such that arrays that help with the
remapping only need to be calculated in one place. Both making the
runtime faster and making it easier to update in the future if
`transform_indep` changes.

And I also added `conf.upstream_inference_transforms` to mirror
`conf.upstream_training_transforms` such that the conditions_dict can
actually be set up different train/inference.