Added flags to prevent writing trajectory but did not change defaults (#276)
This adds these flags:

``` yaml
inference:
  write_trajectory: True # Boolean. Should the /traj/ video pdb file be generated? Takes up a lot of space in production
  write_trajectory_only_t: [] # list(int). Only write trajectory for these given t steps. Example: [50,40,30,20]
  write_trb: True # Boolean. Should the .trb file be written?
  write_trb_indep: True # Boolean. Should the indep be written to the .trb? Adds about 200KB to trb
  write_trb_trajectory: True # Boolean. Should the trajectory be written to the .trb? Adds about 10MB to trb
  ```

At present, to disable trb, you have to turn off all 3. But after next week once the default flips, to enable trb you only have to turn on write_trb.


Try adding this to your url for an easier time viewing the diff because a lot of it is indentation: `?w=1`


This part of the code isn't covered by unit tests (and it would be hard to cover). But I've tested out all the option flag combinations and you always get the same output.