part one of symmetry update (#313)
All the new symmetry manager machinery is present, but is not yet wired
up.

This pr adds the ipd repository as a submodule. The ipd repo is intended
for general purpose code that can potentially be reused in many
projects. For example, most of the non diffusion/rf2aaa specific
symmetry code lives in ipd.

This pr moves project configuration to pyproject.toml.

This pr includes an __init__.py file for rf_diffusion

This pr auto-adds a pre-commit hook that runs "ruff check --fix
rf_diffusion" and does yapf code formatting. Currently code formatting
is disabled on most files, but will be enabled for new files unless
those files are added to the .yapf_exclude file.