Add SSSprinkle Transform to allow on-the-fly generation of simple SS strings (#328)
This PR adds the SSSprinkle inference transform that allows one to
generate simple SS tensor.

The main reason people would want this is to make helical bundles.

Adds this:
```
SSSprinkleTransform:
  active: False
  background: MASK
  helix_chunk_size: 3
  strand_chunk_size: 3
  loop_chunk_size: 3
  mask_chunk_size: 3
  min_helix: 0
  max_helix: 0
  min_strand: 0
  max_strand: 0
  min_loop: 0
  max_loop: 0
  min_mask: 0
  max_mask: 0
  spread_efficiency: 0.7
  chain0_only: True
```


I also fixed a bug in the inference guidepost processing for atomized
residues that's been present (thanks to me) for a few months. Meant to
merge that earlier. But now it's unit tested too.

Also forgot to flip the ts vector in the inference output so people have
been output crazy structures when using `inference.write_extra_ts`