Add hbond_map and target_hbond_satisfaction_cond (#327)
This PR does 3 things:

1. Add `calculate_hbond_map()` based on RifDock's hbond code that has a
0.9 pearson-R to Rosetta
2. Add the target_hbond_satisfaction condition to take advantage of the
new hbonds
3. Add `pop_conditions_dict()` such that conditions dict can be updated
whenever the indep is cropped during training

The `hbond_map` calculation is super fast. At 0.1s for an 800aa input,
this function is basically free. Additionally, the returned vector is
very fast to work with because the "map" is actually just a tensor.

The target satisfaction condition implements the following features
which are pretty self expanatory:
```
 binder_bb_donates_to_target_bb
 binder_bb_accepts_from_target_bb
 binder_sc_donates_to_target_bb
 binder_sc_accepts_from_target_bb
 binder_bb_donates_to_target_sc
 binder_bb_accepts_from_target_sc
 binder_sc_donates_to_target_sc
 binder_sc_accepts_from_target_sc
 binder_bb_donates_to_target_atom
 binder_bb_accepts_from_target_atom
 binder_sc_donates_to_target_atom
 binder_sc_accepts_from_target_atom
 binder_HIS_accepts_from_target_bb
 binder_HIS_accepts_from_target_sc
 binder_HIS_accepts_from_target_atom
 ```
 
Lastly, `pop_conditions_dict()` is a future-facing update for when people start wanting to pull in conditions from datahub. It wasn't necessary for this PR, but since the `conditions_dict['hbond_target_satisfaction']` element has a special `pop_mask()` function, I figured now would be a good time to add it.