PPI RFDiffusion benchmarks (#340)
This one's for Woody.

Reasons why we shouldn't  merge this as-is:

1. The best checkpoint I have has some stuff baked into it that we don't
want. We need to retrain "the checkpoint" before we merge
2. The AF2 initial guess apptainer needs an update so we don't have to
exclude those nodes
3. The wandb thing is uhhhh, we should fix that somehow


Also, the rfdiffusion (old) benchmark currently doesn't finish. It gets
all the way to per_sequence_metrics.py then fails because it can't find
`denoised_xyz_stack`


```
./benchmark/pipeline.py --config-name=ppi_rf_bench 
./benchmark/pipeline.py --config-name=ppi_rf_bench_rfdiffusion
```

---------

Co-authored-by: Woody Ahern <ahern@uw.edu>