Add filters to stop trajectories early (#338)
This PR adds a framework for stopping runs early as well as reporting
values from mid-run.

In doing so, I had to make some decisions.

The filters get defined like this:
```
filters:
    names:
        - MyFilterClass
        - DifferentName:MyFilterClass
    configs:
        MyFilterClass:
            t: '50,49,40,20'
            whatever: 5
        DifferentName:
            t: '15,10,4'
            whatever: 16
```
Filters can choose what t they run at and also accept `suffix` and
`prefix` arguments for their terms in the **scorefile**


**Scorefiles**:
Yes, I had to make scorefiles so that you can actually plan out how to
use your filters. By default (toggleable) scores are written to the .trb
and to a .sc (can be changed to .csv with a flag) file. I think I
handled this pretty well, but it's a change. (Yeah you could just rely
on the .trb but it just seems easier to me if there's also a single
summary file instead of many individual files)

If no filters are enabled, there will not be a scorefile produced


**Checkpointing failed runs**:
The silent file checkpoint system already had a clean way to do this

For normal runs (trb/pdb), I decided to drop an empty .trb for the
design. I can't really think of a better way to do this


**Testing filters**:
This worked out really well and will hopefully ensure compliance with
unit testing


**Preventing model load for totally-checkpointed invocations**:
This one wasn't necessary but I think it's important we do it
eventually. The model shouldn't be loaded (in my opinion) unless we're
sure we're going to use it. The only way I saw to do this was by
creating a sampler and not initializing it. I hopefully did this in the
cleanest way possible such that people won't accidentally grab conf
values that will be overwritten by the model (but this quasi-initialized
state only exists briefly in run_inference.py so unless people are
hacking they'll never see it). Additionally, none of the calls to
sampler() work so people won't get very far.