Refactor the run_inference loop for easier protocols (#345)
This PR refactors a bunch of the run_inference main loop to make it
easier to extend for custom protocols.

TSetups act like transforms on the list of ts. This allows multiple
different protocol steps to co-exist in the same trajectory (and for
there not to be an if cluster)

MidRunModifiers allow virtually any operation to happen inside the
for-loop (but thankfully without if-statements). I replaced all of the
if-statements I've added to that for-loop with these modifiers and it
paves the way for several protocol things I added here as well as Will's
symmetry stuff.